refactor: move to their own module e2store types and file handling
