+++ crates/era/src/e2s/mod.rs
@@ -1,0 +1,5 @@
+//! Core e2store primitives and file handling.
+
+pub mod error;
+pub mod file;
+pub mod types;
