+++ crates/era/src/e2s/types.rs
@@ -8,9 +8,9 @@
 //! An [`Entry`] is a complete record in the file, consisting of both a [`Header`] and its
 //! associated data
 
+use crate::e2s::error::E2sError;
 use ssz_derive::{Decode, Encode};
 use std::io::{self, Read, Write};
-use thiserror::Error;
 
 /// [`Version`] record: ['e', '2']
 pub const VERSION: [u8; 2] = [0x65, 0x32];
@@ -20,34 +20,6 @@ pub const EMPTY: [u8; 2] = [0x00, 0x00];
 
 /// `SlotIndex` record: ['i', '2']
 pub const SLOT_INDEX: [u8; 2] = [0x69, 0x32];
-
-/// Error types for e2s file operations
-#[derive(Error, Debug)]
-pub enum E2sError {
-    /// IO error during file operations
-    #[error("IO error: {0}")]
-    Io(#[from] io::Error),
-
-    /// Error during SSZ encoding/decoding
-    #[error("SSZ error: {0}")]
-    Ssz(String),
-
-    /// Reserved field in header not zero
-    #[error("Reserved field in header not zero")]
-    ReservedNotZero,
-
-    /// Error during snappy compression
-    #[error("Snappy compression error: {0}")]
-    SnappyCompression(String),
-
-    /// Error during snappy decompression
-    #[error("Snappy decompression error: {0}")]
-    SnappyDecompression(String),
-
-    /// Error during RLP encoding/decoding
-    #[error("RLP error: {0}")]
-    Rlp(String),
-}
 
 /// Header for TLV records in e2store files
 #[derive(Debug, Clone, PartialEq, Eq, Encode, Decode)]
