+++ crates/era/tests/it/dd.rs
@@ -4,7 +4,7 @@
 use alloy_consensus::{BlockBody, Header};
 use alloy_primitives::U256;
 use reth_era::{
-    e2s_types::IndexEntry,
+    e2s::types::IndexEntry,
     era1_file::{Era1Reader, Era1Writer},
     era_file_ops::{StreamReader, StreamWriter},
     execution_types::CompressedBody,
