+++ crates/era/tests/it/roundtrip.rs
@@ -10,7 +10,7 @@
 use alloy_consensus::{BlockBody, BlockHeader, Header, ReceiptWithBloom};
 use rand::{prelude::IndexedRandom, rng};
 use reth_era::{
-    e2s_types::IndexEntry,
+    e2s::types::IndexEntry,
     era1_file::{Era1File, Era1Reader, Era1Writer},
     era1_types::{Era1Group, Era1Id},
     era_file_ops::{EraFileFormat, StreamReader, StreamWriter},
